requirements.txt
futurehouse-client==0.3.17
httptools==0.6.4
jwt==1.3.1
langchain==0.3.26
langchain-openai==0.3.25
pypandoc==1.15
setuptools==80.1.0
